Add a specialised Conda package-building image
Add centos-7-conda-build for building Conda packages. This replaces
centos-7-conda as the build image.The latter now hosts only a basic
Conda installation intended to be both the base image on which
centos-7-conda-build is built, and a standalone Conda installation for
runtime deployment of Conda packages, using with docker-compose.

Add images based on bitnami/minideb because it is much lighter than
Ubuntu and is built upstream every 24 hours, making it faster to pick
up security updates.

Add unattended-upgrades to all Ubuntu/Debian builds to get security
updates, rather than relying on upstream updates and then rebuilding
base images. As a result, we can build without --pull. We do this for
bitnami/minideb for consistency, even though their images should be
less than 24 hours old.

Improve locale configuration so that it does not give warnings during
image building.

Update Conda to 4.10.3 and conda-build to 3.21.7.

Remove the Ubuntu 16.04 Conda Dockerfile because we've never used it
to build or run Conda packages.