Make singularity-wrapper more portable
This replaces use of bash in the wrapper generator that is hosted
inside the container, making it usable in containers that do not have
bash installed e.g. Alpine Linux.

It also removes a dependency on jq, again for portability. This
requires a change the manifest of executables from JSON to a flat
list.