Remove irods-dev from server images
As of iRODS 4.3.2, including irods-dev in the default server image
results in the final image being ~4GiB, which is about 4x larger
than iRODS 4.3.1 and too large to be manageable.

For consistency, remove irods-dev from the images of earlier
releases. From this point, use the client-dev image if irods-dev
is required.