Avoid using operations that require root in the Docker entrypoints
In order to align internal and external UID/GIDs on Linux the Docker entrypoints
were doing late user creation, which requires root. This causes problems when
using the resulting images from unprivileged Singularity.

We can avoid using this and get more generally usable images. The original use
case for having UID/GID aligned was for cases when e.g. downloading with iget
into a bind-mounted directory on Linux. Given that using Singularity makes this
trivial, we can go ahead and remove the feature. On macOS, Singularity is not
available, however, Docker aligns the UID/GID for us, so that is not an issue.